+++ ios/airship_flutter/Package.swift
@@ -10,13 +10,15 @@ let package = Package(
         .library(name: "airship-flutter", targets: ["airship_flutter"])
     ],
     dependencies: [
-        .package(url: "https://github.com/urbanairship/airship-mobile-framework-proxy.git", from: "15.16.0")
+        .package(url: "https://github.com/urbanairship/airship-mobile-framework-proxy.git", from: "16.0.0"),
+        .package(url: "https://github.com/urbanairship/ios-library.git", from: "21.0.1")
     ],
     targets: [
         .target(
             name: "airship_flutter",
             dependencies: [
-                .product(name: "AirshipFrameworkProxy", package: "airship-mobile-framework-proxy")
+                .product(name: "AirshipFrameworkProxy", package: "airship-mobile-framework-proxy"),
+                .product(name: "AirshipScenes", package: "ios-library")
             ],
             resources: [
                  .process("PrivacyInfo.xcprivacy")